ENH: sketch ml code

See merge request oddkiva/sara!464

+++ cpp/src/DO/Sara/RANSAC/AContrario.hpp
@@ -66,15 +66,16 @@ namespace DO::Sara {
     return r;
   }
 
-
   // Calculate the Number of False Alarms (NFA).
   template <typename MinimalSolver>
   auto log_nfa(const std::size_t num_data_points, const std::size_t subset_size,
                const double normalized_residual) -> double
   {
     return std::log(MinimalSolver::num_candidate_models) +
-           log_combinatorial(num_data_points, subset_size) +
-           log_combinatorial(subset_size, MinimalSolver::num_points) +
+           log_combinatorial(static_cast<int>(num_data_points),
+                             static_cast<int>(subset_size)) +
+           log_combinatorial(static_cast<int>(subset_size),
+                             MinimalSolver::num_points) +
            std::log(normalized_residual) * subset_size;
   }
